requirements/base.txt
# Core dependencies for the Telegram bot
python-telegram-bot[job-queue]>=20.0
pyairtable>=1.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rapidfuzz>=3.0.0
